Explicitly run scripts through Python
Replace Python script shell invocation and pass the scripts directly to
the Python parser.  This prevents reliance on the /usr/bin/env syntax.
(For example, `python3` is the only available parser in some systems.)

The PYTHON environment variable is also used to permit the environment
to define the Python parser.