Restart setup: Replace -z test with -eq
When determining the runtime from input.nml, the -z test would look for
an empty string.  But this would incorrectly report a valid interval if
the runtime was set to 0, while DAYMAX was also set in MOM_input.

This tries to remedy the issue by replacing -z with -eq.

Now, "" -eq "0" and "0" -eq "0" all report true (as does "abc" -eq "0").
But "1" -eq "0" will report false.